wininternl.h port from wine

+++ include/WinAPI/Win32/ntdef.h
@@ -21,7 +21,7 @@
 
 #include "specstrings.h"
 
-typedef unsigned long* ULONG_PTR;
+typedef unsigned int* ULONG_PTR;
 
 #ifdef __cplusplus
 extern "C" {
